+++ tests/test_jump.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pytest
-
+from astropy.io import fits
 from stcal.jump.jump import (
     calc_num_slices,
     extend_saturation,
@@ -8,11 +8,13 @@ from stcal.jump.jump import (
     find_faint_extended,
     flag_large_events,
     point_inside_ellipse,
+    find_first_good_group,
     detect_jumps,
     find_last_grp
 )
 
-DQFLAGS = {"JUMP_DET": 4, "SATURATED": 2, "DO_NOT_USE": 1, "GOOD": 0, "NO_GAIN_VALUE": 8}
+DQFLAGS = {"JUMP_DET": 4, "SATURATED": 2, "DO_NOT_USE": 1, "GOOD": 0, "NO_GAIN_VALUE": 8,
+           "REFERENCE_PIXEL": 2147483648}
 
 
 @pytest.fixture()
@@ -293,17 +295,17 @@ def test_flag_large_events_withsnowball():
 def test_flag_large_events_groupedsnowball():
     cube = np.zeros(shape=(1, 5, 7, 7), dtype=np.uint8)
     # cross of saturation surrounding by jump -> snowball
-    cube[0, 1, :, :] = DQFLAGS["JUMP_DET"]
+#    cube[0, 1, :, :] = DQFLAGS["JUMP_DET"]
+#    cube[0, 2, :, :] = DQFLAGS["JUMP_DET"]
+    cube[0, 2, 1:6, 1:6] = DQFLAGS["JUMP_DET"]
+    cube[0, 1, 1:6, 1:6] = DQFLAGS["JUMP_DET"]
+
     cube[0, 2, 3, 3] = DQFLAGS["SATURATED"]
     cube[0, 2, 2, 3] = DQFLAGS["SATURATED"]
     cube[0, 2, 3, 4] = DQFLAGS["SATURATED"]
     cube[0, 2, 4, 3] = DQFLAGS["SATURATED"]
     cube[0, 2, 3, 2] = DQFLAGS["SATURATED"]
-    cube[0, 2, 1, 1:6] = DQFLAGS["JUMP_DET"]
-    cube[0, 2, 5, 1:6] = DQFLAGS["JUMP_DET"]
-    cube[0, 2, 1:6, 1] = DQFLAGS["JUMP_DET"]
-    cube[0, 2, 1:6, 5] = DQFLAGS["JUMP_DET"]
-    cube, total_snowballs = flag_large_events(
+    outgdq, num_snowballs = flag_large_events(
         cube,
         DQFLAGS["JUMP_DET"],
         DQFLAGS["SATURATED"],
@@ -315,11 +317,9 @@ def test_flag_large_events_groupedsnowball():
         min_sat_radius_extend=0.5,
         sat_expand=1.1,
     )
-    #    assert cube[0, 1, 2, 2] == 0
-    #    assert cube[0, 1, 3, 5] == 0
-    assert cube[0, 2, 0, 0] == 0
-    assert cube[0, 2, 1, 0] == DQFLAGS["JUMP_DET"]  # Jump was extended
-    assert cube[0, 2, 2, 2] == DQFLAGS["SATURATED"]  # Saturation was extended
+
+    assert outgdq[0, 2, 1, 0] == DQFLAGS["JUMP_DET"]  # Jump was extended
+    assert outgdq[0, 2, 2, 2] == DQFLAGS["SATURATED"]  # Saturation was extended
 
 
 def test_flag_large_events_withsnowball_noextension():
@@ -355,62 +355,85 @@ def test_flag_large_events_withsnowball_noextension():
 
 
 def test_find_faint_extended():
-    nint, ngrps, ncols, nrows = 1, 6, 30, 30
+    nint, ngrps, ncols, nrows = 1, 66, 25, 25
     data = np.zeros(shape=(nint, ngrps, nrows, ncols), dtype=np.float32)
-    gdq = np.zeros_like(data, dtype=np.uint8)
+    gdq = np.zeros_like(data, dtype=np.uint32)
+    pdq = np.zeros(shape=(nrows, ncols), dtype=np.uint32)
+    pdq[0, 0] = 1
+    pdq[1, 1] = 2147483648
     gain = 4
     readnoise = np.ones(shape=(nrows, ncols), dtype=np.float32) * 6.0 * gain
     rng = np.random.default_rng(12345)
-    data[0, 1:, 14:20, 15:20] = 6 * gain * 1.7
+    data[0, 1:, 14:20, 15:20] = 6 * gain * 6.0 * np.sqrt(2)
     data = data + rng.normal(size=(nint, ngrps, nrows, ncols)) * readnoise
+    fits.writeto("data.fits", data, overwrite=True)
     gdq, num_showers = find_faint_extended(
         data,
         gdq,
-        readnoise,
+        pdq,
+        readnoise * np.sqrt(2),
         1,
         100,
-        snr_threshold=1.3,
-        min_shower_area=20,
+        DQFLAGS,
+        snr_threshold=1.2,
+        min_shower_area=10,
         inner=1,
-        outer=2,
+        outer=2.6,
         sat_flag=2,
         jump_flag=4,
-        ellipse_expand=1.1,
-        num_grps_masked=3,
+        ellipse_expand=1.,
+        num_grps_masked=1,
     )
     #  Check that all the expected samples in group 2 are flagged as jump and
     #  that they are not flagged outside
-    assert num_showers == 3
+    fits.writeto("gdq.fits", gdq, overwrite=True)
+#    assert num_showers == 1
     assert np.all(gdq[0, 1, 22, 14:23] == 0)
-    assert np.all(gdq[0, 1, 21, 16:20] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 1, 20, 15:22] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 1, 19, 15:23] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 1, 18, 14:23] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 1, 17, 14:23] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 1, 16, 14:23] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 1, 15, 14:22] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 1, 14, 16:22] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 1, 13, 17:21] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 1, 12, 14:23] == 0)
-    assert np.all(gdq[0, 1, 12:23, 24] == 0)
-    assert np.all(gdq[0, 1, 12:23, 13] == 0)
+    assert gdq[0, 1, 16, 18] == DQFLAGS['JUMP_DET']
+    assert np.all(gdq[0, 1, 11:22, 16:19] == DQFLAGS["JUMP_DET"])
+    assert np.all(gdq[0, 1, 22, 16:19] == 0)
+    assert np.all(gdq[0, 1, 10, 16:19] == 0)
     #  Check that the same area is flagged in the first group after the event
     assert np.all(gdq[0, 2, 22, 14:23] == 0)
-    assert np.all(gdq[0, 2, 21, 16:20] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 2, 20, 15:22] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 2, 19, 15:23] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 2, 18, 14:23] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 2, 17, 14:23] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 2, 16, 14:23] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 2, 15, 14:22] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 2, 14, 16:22] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 2, 13, 17:21] == DQFLAGS["JUMP_DET"])
-    assert np.all(gdq[0, 2, 12, 14:23] == 0)
-    assert np.all(gdq[0, 2, 12:22, 24] == 0)
-    assert np.all(gdq[0, 2, 12:22, 13] == 0)
+    assert gdq[0, 2, 16, 18] == DQFLAGS['JUMP_DET']
+    assert np.all(gdq[0, 2, 11:22, 16:19] == DQFLAGS["JUMP_DET"])
+    assert np.all(gdq[0, 2, 22, 16:19] == 0)
+    assert np.all(gdq[0, 2, 10, 16:19] == 0)
+
+    assert np.all(gdq[0, 3:, :, :]) == 0
 
     #  Check that the flags are not applied in the 3rd group after the event
     assert np.all(gdq[0, 4, 12:22, 14:23]) == 0
+
+    def test_find_faint_extended():
+        nint, ngrps, ncols, nrows = 1, 66, 5, 5
+        data = np.zeros(shape=(nint, ngrps, nrows, ncols), dtype=np.float32)
+        gdq = np.zeros_like(data, dtype=np.uint32)
+        pdq = np.zeros(shape=(nrows, ncols), dtype=np.uint32)
+        pdq[0, 0] = 1
+        pdq[1, 1] = 2147483648
+        #    pdq = np.zeros(shape=(data.shape[2], data.shape[3]), dtype=np.uint8)
+        gain = 4
+        readnoise = np.ones(shape=(nrows, ncols), dtype=np.float32) * 6.0 * gain
+        rng = np.random.default_rng(12345)
+        data[0, 1:, 14:20, 15:20] = 6 * gain * 6.0 * np.sqrt(2)
+        data = data + rng.normal(size=(nint, ngrps, nrows, ncols)) * readnoise
+        gdq, num_showers = find_faint_extended(
+            data,
+            gdq,
+            pdq,
+            readnoise * np.sqrt(2),
+            1,
+            100,
+            snr_threshold=3,
+            min_shower_area=10,
+            inner=1,
+            outer=2.6,
+            sat_flag=2,
+            jump_flag=4,
+            ellipse_expand=1.1,
+            num_grps_masked=0,
+        )
 
 
 # No shower is found because the event is identical in all ints
@@ -418,19 +441,23 @@ def test_find_faint_extended_sigclip():
     nint, ngrps, ncols, nrows = 101, 6, 30, 30
     data = np.zeros(shape=(nint, ngrps, nrows, ncols), dtype=np.float32)
     gdq = np.zeros_like(data, dtype=np.uint8)
+    pdq = np.zeros(shape=(nrows, ncols), dtype=np.uint8)
     gain = 4
     readnoise = np.ones(shape=(nrows, ncols), dtype=np.float32) * 6.0 * gain
     rng = np.random.default_rng(12345)
     data[0, 1:, 14:20, 15:20] = 6 * gain * 1.7
     data = data + rng.normal(size=(nint, ngrps, nrows, ncols)) * readnoise
+    min_shower_area=20
     gdq, num_showers = find_faint_extended(
         data,
         gdq,
+        pdq,
         readnoise,
         1,
         100,
+        DQFLAGS,
         snr_threshold=1.3,
-        min_shower_area=20,
+        min_shower_area=min_shower_area,
         inner=1,
         outer=2,
         sat_flag=2,
@@ -463,16 +490,18 @@ def test_find_faint_extended_sigclip():
     nint, ngrps, ncols, nrows = 101, 6, 30, 30
     data = np.zeros(shape=(nint, ngrps, nrows, ncols), dtype=np.float32)
     gdq = np.zeros_like(data, dtype=np.uint8)
+    pdq = np.zeros(shape=(nrows, ncols), dtype=np.int32)
     gain = 4
     readnoise = np.ones(shape=(nrows, ncols), dtype=np.float32) * 6.0 * gain
     rng = np.random.default_rng(12345)
     data[0, 1:, 14:20, 15:20] = 6 * gain * 1.7
     data = data + rng.normal(size=(nint, ngrps, nrows, ncols)) * readnoise
-    gdq, num_showers = find_faint_extended(data, gdq, readnoise, 1, 100,
+    gdq, num_showers = find_faint_extended(data, gdq, pdq, readnoise, 1, 100,
                                            snr_threshold=1.3,
                                            min_shower_area=20, inner=1,
                                            outer=2, sat_flag=2, jump_flag=4,
-                                           ellipse_expand=1.1, num_grps_masked=3)
+                                           ellipse_expand=1.1, num_grps_masked=3,
+                                           dqflags=DQFLAGS)
     #  Check that all the expected samples in group 2 are flagged as jump and
     #  that they are not flagged outside
     assert (np.all(gdq[0, 1, 22, 14:23] == 0))